requirements.txt
# Amazon EC Tool - Requirements
# 基本ライブラリ
requests>=2.28.0
boto3>=1.26.0
pandas>=1.5.0
numpy>=1.24.0

# Google API関連
google-api-python-client>=2.0.0
gspread>=5.7.0
oauth2client>=4.1.3

# 設定・ユーティリティ
PyYAML>=6.0
python-dotenv>=0.21.0
schedule>=1.2.0

# セキュリティ・暗号化
cryptography>=39.0.0

# 開発・テスト用
pytest>=7.3.0
black>=23.0.0
flake8>=6.0.0 